requirements.txt
streamlit
transformers
torch
accelerate
python-docx